refactor(tests): extract duplicated git-fixture helpers to GitFixtureRunner
`RunGit` + `StageFixtureBaseline` (plus the trivial `StageAndCommitAll`
wrapper) were duplicated verbatim across `ValidateRecentGitChangesTests` and
`ValidateWorkspaceChangeTrackerReconcileTests`. Promote both copies into a
single `GitFixtureRunner` static under `tests/RoslynMcp.Tests/Support/`.

`RunGitCapture` (only in ValidateWorkspaceChangeTrackerReconcileTests) stays
local — it has a different stdout-capture shape and is single-use.

Closes backlog row `test-git-fixture-helper-dedup`.
