+++ ai_docs/backlog.md
@@ -3,7 +3,7 @@
 <!-- purpose: Open work only; contract for agents syncing backlog on ship. -->
 <!-- scope: in-repo -->
 
-**updated_at: 2026-05-26T03:00:00Z**
+**updated_at: 2026-05-26T04:00:00Z**
 
 ## Agent contract
 
@@ -64,7 +64,6 @@
 | id | pri | deps | do |
 |----|-----|------|-----|
 | `tool-surface-pagination-or-tool-sets` | Low | none | 171 tools is approaching small-model discovery saturation, but current evidence points to a routing/steering problem more than a raw-count problem. Now that `recommend_workflow` exists, wait for fresh post-router evidence before adding tool-set catalog resources (for example `roslyn://server/catalog/tool-sets` and `roslyn://server/catalog/tools/{toolSet}/{offset}/{limit}`) that expose bounded subsets such as `navigation`, `refactoring`, `validation`, and `analysis` without hiding tools from clients that can handle the full surface. Do not change MCP tool registration or default `tools/list` visibility in this row unless the implementation note proves the client supports that safely. Anchors: `src/RoslynMcp.Host.Stdio/Catalog/ServerSurfaceCatalog.cs`, `src/RoslynMcp.Host.Stdio/Resources/ServerResources.cs`, `tests/RoslynMcp.Tests/SurfaceCatalogTests.cs`. Regression test shape: catalog summary advertises the tool-set resource; each named set returns only matching categories with offset/limit/hasMore metadata; unknown set returns structured `InvalidArgument`; existing full and paginated catalog resources remain unchanged. Evidence: surface count drift across recent versions (`server_info.surface.registered.tools` 151 -> 171 since v1.27); MCP spec tools/list pagination; 2026-05-20 no-context subagent probe showed correctly steered agents chose Roslyn primitives without needing a full catalog dump. **Weaker evidence - N until small-model discovery friction is reported after the router lands externally.** Source: 2026-05-05 MCP-best-practices comparison §3 rec J plus 2026-05-20 agent-view review. |
-| `test-git-fixture-helper-dedup` | Low | none | `RunGit` + `StageFixtureBaseline` helpers are duplicated verbatim (similarity=1.0, 30+38 lines) across `ValidateRecentGitChangesTests` and `ValidateWorkspaceChangeTrackerReconcileTests`. Extract a `GitFixtureRunner` static under `tests/RoslynMcp.Tests/Support/`; both test classes delegate. Anchors: `tests/RoslynMcp.Tests/ValidateRecentGitChangesTests.cs:257-325`; `tests/RoslynMcp.Tests/ValidateWorkspaceChangeTrackerReconcileTests.cs:280-348`; new `tests/RoslynMcp.Tests/Support/GitFixtureRunner.cs`. Regression test shape: both originating test classes stay green after delegation. Evidence: `mcp__roslyn__find_duplicated_methods` normalizedHash 5157e720eefee8e1 + d4c1d337fbd2629d. Source: 2026-05-26 discovery-sweep refactor audit. |
 | `brainstorm-report-refresh-shipped-overlaps` | Low | none | `audit-reports/application-brainstorm.md` cites 4 "exact open" backlog rows that are stale: BRAIN-001 cites shipped `workspace-fork-apply-primitive` (the `workspace_fork_apply` tool is in `ValidationBundleTools.cs`); BRAIN-003 cites shipped `plugin-package-files-allowlist` (per `changelog.d/workspace-fork-apply-tool.md`); BRAIN-004 cites absent `surface-test-resumability-cleanup-skill`; BRAIN-006 cites absent `initiative-executor-roslyn-tool-discovery-brief`. Mark BRAIN-001/003 as `implemented`; re-derive BRAIN-004/006 against the current backlog; refresh the Status / Backlog-overlap section. Anchors: `audit-reports/application-brainstorm.md:28-126`. Evidence: current backlog grep for those row IDs returns none. Source: 2026-05-26 discovery-sweep work-search. |
 
 ## Defer
